Separate uri.rs into module folder
